Move parsing functions into seperate file
