+++ scripts/plotting/fakes/smoothing_parameters.py
@@ -2,7 +2,7 @@ import matplotlib as mpl
 import numpy as np
 
 from utilities import boostHistHelpers as hh
-from utilities import common, logging
+from utilities import logging, parsing
 from utilities.io_tools import output_tools
 from utilities.styles import styles
 from wremnants import histselections as sel
@@ -104,7 +104,7 @@ def plot_params(h, params, params_err, label=None, suffix="", proc=""):
 
 
 if __name__ == "__main__":
-    parser = common.plot_parser()
+    parser = parsing.plot_parser()
     parser.add_argument(
         "infile",
         help="Output file of the analysis stage, containing ND boost histograms",
